+++ tests/estimators/object_detection/test_pytorch_object_detector.py
@@ -27,106 +27,6 @@ from tests.utils import ARTTestException
 logger = logging.getLogger(__name__)
 
 
-@pytest.fixture()
-def get_pytorch_object_detector(get_default_mnist_subset):
-    """
-    This class tests the PyTorchObjectDetector object detector.
-    """
-    import torch
-    import torchvision
-
-    from art.estimators.object_detection.pytorch_object_detector import PyTorchObjectDetector
-
-    # Define object detector
-    model = torchvision.models.detection.fasterrcnn_resnet50_fpn(
-        pretrained=True, progress=True, num_classes=91, pretrained_backbone=True
-    )
-    params = [p for p in model.parameters() if p.requires_grad]
-    optimizer = torch.optim.SGD(params, lr=0.01)
-
-    object_detector = PyTorchObjectDetector(
-        model=model,
-        optimizer=optimizer,
-        clip_values=(0, 1),
-        channels_first=False,
-        attack_losses=["loss_classifier", "loss_box_reg", "loss_objectness", "loss_rpn_box_reg"],
-    )
-
-    (_, _), (x_test_mnist, _) = get_default_mnist_subset
-
-    x_test = np.transpose(x_test_mnist[:2], (0, 2, 3, 1))
-    x_test = np.repeat(x_test.astype(np.float32), repeats=3, axis=3)
-
-    # Create labels
-    result = object_detector.predict(x=x_test)
-
-    y_test = [
-        {
-            "boxes": result[0]["boxes"],
-            "labels": result[0]["labels"],
-            "scores": np.ones_like(result[0]["labels"]),
-        },
-        {
-            "boxes": result[1]["boxes"],
-            "labels": result[1]["labels"],
-            "scores": np.ones_like(result[1]["labels"]),
-        },
-    ]
-
-    yield object_detector, x_test, y_test
-
-
-@pytest.fixture()
-def get_pytorch_object_detector_mask(get_default_mnist_subset):
-    """
-    This class tests the PyTorchObjectDetector object detector.
-    """
-    import torch
-    import torchvision
-
-    from art.estimators.object_detection.pytorch_object_detector import PyTorchObjectDetector
-
-    # Define object detector
-    model = torchvision.models.detection.maskrcnn_resnet50_fpn(
-        pretrained=True, progress=True, num_classes=91, pretrained_backbone=True
-    )
-    params = [p for p in model.parameters() if p.requires_grad]
-    optimizer = torch.optim.SGD(params, lr=0.01)
-
-    object_detector = PyTorchObjectDetector(
-        model=model,
-        optimizer=optimizer,
-        clip_values=(0, 1),
-        channels_first=False,
-        attack_losses=["loss_classifier", "loss_box_reg", "loss_objectness", "loss_rpn_box_reg"],
-    )
-
-    (_, _), (x_test_mnist, _) = get_default_mnist_subset
-
-    x_test = np.transpose(x_test_mnist[:2], (0, 2, 3, 1))
-    x_test = np.repeat(x_test.astype(np.float32), repeats=3, axis=3)
-
-    # Create labels
-    result = object_detector.predict(x=x_test)
-
-    y_test = [
-        {
-            "boxes": result[0]["boxes"],
-            "labels": result[0]["labels"],
-            "scores": np.ones_like(result[0]["labels"]),
-            "masks": result[0]["masks"],
-        },
-        {
-            "boxes": result[1]["boxes"],
-            "labels": result[1]["labels"],
-            "scores": np.ones_like(result[1]["labels"]),
-            "masks": result[0]["masks"],
-        },
-    ]
-
-    yield object_detector, x_test, y_test
-
-
 @pytest.mark.only_with_platform("pytorch")
 def test_predict(art_warning, get_pytorch_object_detector):
     try:
